IMPORTANT UPDATE!!!!
I refactored the way that the page builder and page object work to a certain degree. I set it up so you set the builder file in config. This must be added to your config file:

<builder>
	<stack>default.xml</stack>
	<persistPage>true</persistPage>
</builder>

This adds a great deal of flexibility to the way the pages are constructed, and enables you to use the builder outside the scope of the current page; now you can use it to build any page you want.